+++ web/src/components/molecules/Request/Details/SidebarWrapper.tsx
@@ -54,7 +54,9 @@ const RequestSidebarWrapper: React.FC<Props> = ({
 
   const handleSubmit: FocusEventHandler<HTMLElement> | undefined = useCallback(async () => {
     const requestId = currentRequest?.id;
-    if (!requestId || selectedReviewers.length === 0) {
+    const isEqual =
+      JSON.stringify([...defaultValue].sort()) === JSON.stringify([...selectedReviewers].sort());
+    if (!requestId || selectedReviewers.length === 0 || isEqual) {
       hideViewReviewers();
       return;
     }
@@ -77,6 +79,7 @@ const RequestSidebarWrapper: React.FC<Props> = ({
     currentRequest?.id,
     currentRequest?.state,
     currentRequest?.title,
+    defaultValue,
     hideViewReviewers,
     onRequestUpdate,
     selectedReviewers,
